+++ beacon-node-api/src/main/kotlin/org/web3j/eth2/api/schema/Slot.kt
@@ -15,4 +15,4 @@ package org.web3j.eth2.api.schema
 /**
  * A slot number.
  */
-typealias Slot = Int // uint64
+typealias Slot = Uint64
